+++ api/config.py
@@ -11,5 +11,5 @@ matplotlib.use("agg")
 load_dotenv()
 
 WHITELISTED_CORS_URLS = os.environ.get("WHITELISTED_CORS_URLS", "")
-DEBUG_MODE = os.environ.get("DEBUG_MODE", False)
+DEBUG_MODE = os.environ.get("DEBUG_MODE", False) == "True"
 BASE_PATH = os.environ.get("BASE_PATH", "")