+++ api/exceptions.py
@@ -45,7 +45,7 @@ class InvalidUrlParameterException(HTTPException):
     """
 
     def __init__(self) -> None:
-        super().__init__(status_code=422, detail="Invalid content_url parameter in request.")
+        super().__init__(status_code=422, detail="Invalid content_url parameter in request")
 
 
 # Electrophysiology
@@ -90,7 +90,7 @@ class NoUnitFound(HTTPException):
     "Thrown when no unit is found."
 
     def __init__(self):
-        super().__init__(status_code=404, detail="The NWB file didn't contain any Ic data.")
+        super().__init__(status_code=404, detail="The NWB file didn't contain a 'unit'.")
 
 
 class NoRateFound(HTTPException):