+++ api/router/generate.py
@@ -10,7 +10,7 @@ from fastapi.security import HTTPBearer
 from api.services.trace_img import generate_electrophysiology_image
 from api.services.morpho_img import generate_morphology_image
 from api.dependencies import retrieve_user
-from api.models.common import ImageGenerationInput
+from api.models.common import ErrorMessage, ImageGenerationInput
 from api.user import User
 
 
@@ -21,6 +21,7 @@ require_bearer = HTTPBearer()
 @router.get(
     "/morphology-image",
     dependencies=[Depends(require_bearer)],
+    responses={404: {"model": ErrorMessage}, 422: {"model": ErrorMessage}},
     response_model=None,
 )
 def get_morphology_image(
@@ -41,6 +42,7 @@ def get_morphology_image(
 @router.get(
     "/trace-image",
     dependencies=[Depends(require_bearer)],
+    responses={404: {"model": ErrorMessage}},
     response_model=None,
 )
 def get_trace_image(image_input: ImageGenerationInput = Depends(), user: User = Depends(retrieve_user)) -> Response: