+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
 - Exception handling in API
 - Move functions to `services` and `utils` modules to differentiate business logic
 - Pydoc strings to make them uniform
+- Add unit tests related to `services` and `utils`
+- Replace `matplotlib` `set_tight_layout()` with `set_layout_engine` due to deprecation
 
 ## [0.4.4] - 11/06/2024